+++ bitbox/src/beatree/bbn.rs
@@ -93,6 +93,6 @@ pub struct BbnStoreCommitOutput {
     pub bbn: Vec<BranchNode>,
     pub free_list_pages: Vec<(PageNumber, Box<Page>)>,
     pub bump: PageNumber,
-    pub extend_file_sz: Option<usize>,
+    pub extend_file_sz: Option<u64>,
     pub freelist_head: PageNumber,
 }
